Fixing warnings from the new tests

+++ test/diff2.cpp
@@ -73,7 +73,7 @@ void diff2Test(string pTestFile, bool isSubRef=false, const vector<af_seq> *seqv
     ASSERT_EQ(AF_SUCCESS, af_get_data_ptr((void*)outData, outArray));
 
     // Compare result
-    for (int testIter = 0; testIter < tests.size(); ++testIter) {
+    for (size_t testIter = 0; testIter < tests.size(); ++testIter) {
         vector<T> currGoldBar = tests[testIter];
         size_t nElems = currGoldBar.size();
         for (size_t elIter = 0; elIter < nElems; ++elIter) {
